# src/problem5/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce API
  version: 1.0.0
  description: A RESTful API for e-commerce operations - managing users, products, and orders
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /users:
    get:
      summary: List users
      description: Returns a paginated list of users with optional filtering
      tags:
        - Users
      parameters:
        - name: email
          in: query
          schema:
            type: string
          description: Filter by email (case-insensitive partial match)
        - name: name
          in: query
          schema:
            type: string
          description: Filter by name (case-insensitive partial match)
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, admin]
          description: Filter by user role
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, email, createdAt, updatedAt]
            default: createdAt
          description: Field to sort by
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Maximum number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserList'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a user
      description: Creates a new user and returns it
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created successfully
          headers:
            Location:
              description: URL of the created user
              schema:
                type: string
                example: /users/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /products:
    get:
      summary: List products
      description: Returns a paginated list of products with optional filtering
      tags:
        - Products
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Filter by name (case-insensitive partial match)
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
          description: Filter by product status
        - name: minPrice
          in: query
          schema:
            type: number
            format: float
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
            format: float
          description: Maximum price filter
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, price, stock, createdAt, updatedAt]
            default: createdAt
          description: Field to sort by
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Maximum number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductList'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a product
      description: Creates a new product and returns it
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Product created successfully
          headers:
            Location:
              description: URL of the created product
              schema:
                type: string
                example: /products/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Product ID

    get:
      summary: Get a product by ID
      description: Returns a single product by its ID
      tags:
        - Products
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update a product
      description: Updates only the provided fields of an existing product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProduct'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a product
      description: Permanently deletes a product
      tags:
        - Products
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Orders endpoints
  /orders:
    get:
      summary: List orders
      description: Returns a paginated list of orders with optional filtering
      tags:
        - Orders
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
          description: Filter by user ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered, cancelled]
          description: Filter by order status
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [totalAmount, createdAt, updatedAt]
            default: createdAt
          description: Field to sort by
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
          description: Maximum number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip
      responses:
        '200':
          description: Paginated list of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderList'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create an order
      description: |
        Creates a new order. Validates that:
        - The user exists
        - All products exist and are active
        - Sufficient stock is available for all items

        Stock is automatically reduced for each product when the order is created.
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              description: URL of the created order
              schema:
                type: string
                example: /orders/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error (insufficient stock, inactive product, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Order ID

    get:
      summary: Get an order by ID
      description: Returns a single order by its ID, including all order items
      tags:
        - Orders
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update an order
      description: |
        Updates an order's status or shipping address.

        Status transitions are validated:
        - pending → confirmed, cancelled
        - confirmed → shipped, cancelled
        - shipped → delivered
        - delivered → (no transitions allowed)
        - cancelled → (no transitions allowed)

        When an order is cancelled, stock is automatically restored for all items.
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrder'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error (invalid status transition)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete an order
      description: |
        Permanently deletes an order. Only pending orders can be deleted.
        Stock is automatically restored for all items when the order is deleted.
      tags:
        - Orders
      responses:
        '204':
          description: Order deleted successfully
        '400':
          description: Order cannot be deleted (not in pending status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        email:
          type: string
          format: email
          description: User email address (unique)
          example: "john@example.com"
        name:
          type: string
          description: User's full name
          example: "John Doe"
        role:
          type: string
          enum: [customer, admin]
          description: User role
          example: "customer"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
      required:
        - id
        - email
        - name
        - role
        - createdAt
        - updatedAt
    CreateUser:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address (must be unique)
          example: "john@example.com"
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: User's full name
          example: "John Doe"
        role:
          type: string
          enum: [customer, admin]
          default: customer
          description: User role
          example: "customer"
      required:
        - email
        - name
...
    PaginatedUserList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - data
        - pagination

    # Product schemas
    Product:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        name:
          type: string
          description: Product name
          example: "Wireless Keyboard"
        description:
          type: string
          description: Product description
          example: "A high-quality wireless keyboard with RGB lighting"
        price:
          type: string
          description: Product price (decimal as string for precision)
          example: "99.99"
        stock:
          type: integer
          description: Available stock quantity
          example: 50
        status:
          type: string
          enum: [draft, active, archived]
          description: Product status
          example: "active"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
      required:
        - id
        - name
        - description
        - price
        - stock
        - status
        - createdAt
        - updatedAt
    CreateProduct:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
          example: "Wireless Keyboard"
        description:
          type: string
          maxLength: 2000
          description: Product description
          example: "A high-quality wireless keyboard with RGB lighting"
        price:
          type: number
          format: float
          minimum: 0
          description: Product price
          example: 99.99
        stock:
          type: integer
          minimum: 0
          default: 0
          description: Initial stock quantity
          example: 50
        status:
          type: string
          enum: [draft, active, archived]
          default: draft
          description: Product status
          example: "active"
      required:
        - name
        - price
    UpdateProduct:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
          example: "Updated Keyboard"
        description:
          type: string
          maxLength: 2000
          description: Product description
          example: "Updated description"
        price:
          type: number
          format: float
          minimum: 0
          description: Product price
          example: 89.99
        stock:
          type: integer
          minimum: 0
          description: Stock quantity
          example: 100
        status:
          type: string
          enum: [draft, active, archived]
          description: Product status
          example: "archived"
    PaginatedProductList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - data
        - pagination

    # Order schemas
    Order:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        userId:
          type: integer
          description: ID of the user who placed the order
          example: 1
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
          description: Order status
          example: "pending"
        totalAmount:
          type: string
          description: Total order amount (decimal as string for precision)
          example: "199.98"
        shippingAddress:
          type: string
          description: Shipping address
          example: "123 Main St, City, Country 12345"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          description: List of items in the order
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (ISO 8601)
          example: "2024-01-15T10:30:00.000Z"
      required:
        - id
        - userId
        - status
        - totalAmount
        - shippingAddress
        - items
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          description: Order item ID
          example: 1
        productId:
          type: integer
          description: Product ID
          example: 1
        quantity:
          type: integer
          description: Quantity ordered
          example: 2
        unitPrice:
          type: string
          description: Price per unit at time of order (decimal as string)
          example: "99.99"
      required:
        - id
        - productId
        - quantity
        - unitPrice
    CreateOrder:
      type: object
      properties:
        userId:
          type: integer
          minimum: 1
          description: ID of the user placing the order
          example: 1
        shippingAddress:
          type: string
          minLength: 1
          maxLength: 500
          description: Shipping address
          example: "123 Main St, City, Country 12345"
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              productId:
                type: integer
                minimum: 1
                description: Product ID
                example: 1
              quantity:
                type: integer
                minimum: 1
                description: Quantity to order
                example: 2
            required:
              - productId
              - quantity
          description: List of items to order
      required:
        - userId
        - shippingAddress
        - items
    UpdateOrder:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
          description: New order status (must follow valid transitions)
          example: "confirmed"
        shippingAddress:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated shipping address
          example: "456 New St, City, Country 67890"
    PaginatedOrderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
        - data
        - pagination

    # Shared schemas
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of matching items
          example: 150
        limit:
          type: integer
          description: Maximum results per page
          example: 100
        offset:
          type: integer
          description: Number of results skipped
          example: 0
        hasMore:
          type: boolean
          description: Whether more results are available
          example: true
      required:
        - total
        - limit
        - offset
        - hasMore
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
      required:
        - error